requirements.txt
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-Script==2.0.6
flask-shell-ipython==0.4.0
Flask-SQLAlchemy==2.3.2
marshmallow==2.19.1
PyJWT==1.7.1
PyMySQL==0.9.3
redis==3.2.1
SQLAlchemy==1.3.1
Werkzeug==0.15.1
